services/api_gateway/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
redis[hiredis]==5.2.1
faster-whisper==1.1.0
pyannote.audio==3.3.2
